Add the "events" command
The command allows to check for regular expression patterns in logfiles.
The logfiles can be the output of "drbdsetup events".  Both patterns
which must be present and patterns which must not be present can be
specified, as well as a timeout.  In addition, patterns can be checked
for incrementally from the last match.