test-initial-sync: Wait for the end of the resync
The test was incorrectly checking for "peer-device disk:UpToDate" events
on the resync targets, but this only indicates that they know that the
sync source is up to date; actually, the test should have been for
"device disk:UpToDate".

In addition, make sure that all devices are inconsistent before the
first node is turned into the primary node, and don't limit the resync
bandwidth to make the test finish faster.  (We don't simulate any
application I/O in this test.)