Add --no-rmmod option to prevent lockdep lock class exhaustion
When lockdep is in use, repeatedly loading and unloading the drbd kernel
module eventually leads to the exhaustion of lock classes, which
is logged and causes lockdep to be turned off.  Add a --no-rmmod option
which prevents the test suite from removing the drbd kernel module so
that lots of tests can be run using lockdep.