test-add-volume: Fix for DRBD 8
In DRBD 8, the role (primary / secondary) is an attribute of device
objects.  The role of the entire resource is determined by the role of
each of its devices; changing the role of the resource with drbdadm
changes the role of each of its devices.

In DRBD 9, the role has become an attribute of the resource itself.

This test was previously changing the role of the first node to Primary
with volume 0 alone enabled, then back to Secondary after enabling
volume 1. This was causing a "drbdsetup secondary 1" call on DRBD 8 with
an undefined volume 1.