Add job files and a script for running them
Different test scripts are supposed to test different test scenarios.
Some of them require a fixed number of nodes to run on (for example,
only a single node or exactly two nodes); others can be run on a varying
numbers of nodes (for example, on two or more nodes).

This commit introduces job files; so far, each job file defines which
test script to run on how many nodes.  This allows to define various
"reasonable" test loads easily without hard coding a specific "standard
test load" into the test scripts themselves.

So far, a job file defines how many nodes it needs, and which test
script to run.  Other parameters, such as how much disk space a test
requires, could be added in the future.

In addition, this commit adds a script to run a set of jobs on a set of
nodes.  For example, all jobs in the jobs/ directory can be run on nodes
a, b, and c as follows:

  lib/run-jobs [-v] [-d] a b c -- jobs/*