Fix diskless nodes in Resource.initial_resync()
It needs to wait until all diskless nodes see all volumes on all
diskfull nodes as UpToDate and until diskfull nodes see all volumes
on all other diskfull nodes as UpToDate.

So far it wanted to see all volumes on all nodes as UpToDate, but
that was wrong for diskless nodes.

The first time I fixed this it was in the diskles test.
It was removed by commit 61b4a072, probably because
the author of that commit did not understood what he removed.

I HOPE IT SURVIVES THIS TIME!