Fix the way the testsuite waits until all connections are established
When waiting for events that can happen in arbitrary order the idea
is to build a container object that contains all the objects and then
call the event() method on the container.

While doing so it turned our that we want to wait until the peer-disks
are visible.

That in turn is only possible when we are not waiting for the connect
event. The reason is that peer-device events usually are interleaved
between the connect events. With waiting for all connect events first
we miss the peer-device events in between!!