fix python/syslog.py to NOT close connection for every line
One "request" to handle for SocketServer.TCPServer is basically the
result of one accept(), and once the handle() routine returns,
that connection is closed.

If the syslog client needs to keep doing TCP handshakes for almost every
logline, that's a lot of overhead, and results in lost loglines.

Use the ThreadingTCPServer as base class,
so we get one thread per incoming stream,
and keep draining the stream in our handle() method
until the source stops sending.