fix ctypes in misaligned_bio.py for newer python
Apparently, older python ctypes was ok with our "pointer arithmetic",
newer python ctypes requires explicit ctypes.cast()

Or maybe it was always broken, and we only noticed now,
because it started to segfault finally?

While at it, add an implicit verify run,
double check return values of read and write,
and raise exceptions when expectations are not met.

Also, make it python3 compatible.