New test that shows how live migration was broken up to 9.0.14
Simulate a live migration (kvm) from a diskfull to a diskless node

And between two diskless nodes.

In both cases, drbd had a bug to show the diskful node as
Outdated or Consistent on the diskless node. (After receiving
the first state update from the diskful node after the migration)

Ensuring this comes never back!