new test: resync
This tests generates as many data-generations as you pass nodes, by
disconnecting each node and writing on one in between the disconnect
operations. The "one" node is the "writing_n" in the code.

Then beginning with the oldest the nodes are formed to a chain,
verifying in each step the resync (direction and amount).

Finally the node with the latest data-set is connected to the node
that had the oldest data-set, causing a wave of resyncs propagating
through the chain. That is verified as well.

Finally the chain is formed to a ring and it is verified that
no resync happens at that stage.