Add testing isolation of primary to quorum test
The major use case of isolating an active primary was missed so far.
When the connections to the rest of the cluster are interrupted it
is done in a way that one goes away by closing the connection
to the primary, i.e. the primary recognises it immediately. The others
go away silently.
Then a write is issued and the primary has to recognise that the
other nodes disappeared in the meantime and is not allowed to
create a new current-uuid in that moment.