New test resync-source-becomes-target
This test creats three different versions of the data. It let
the two older nodes begin a resync. When they are in the middle
of the resync operation, the node with the newest data connects
to the resync source of the already running resync.
It has to pause the running resync, obtain the newest data, and
then continue the resync to the downstream peer.

In variant two it starts the upstream resync after downstream
completed