make two tests more efficient
The "add-connect-delete" and "add-path-multiple-times" tests waste a lot
of time simply waiting.

We don't actually have to try three times to make drbd misbehave. If it
does, it will do so on the first try. The sleeps between the statements
and at the beginning of the test are also superfluous, they just
drag out the execution time.

Also swap out the sleep at the end of the test (because "perhaps we'll
still crash") for an actual functionality check. If the VM actually
crashes because of the call, the test will still fail.

This should still verify the same behaviour but drives down execution
time from around 17 seconds to ~1 second.