New test to trigger a wrong state transfer in DRBD
When DRBD exchanges UUIDs after establishing a conneciton. It sends its
current UUID and expects to learn about the current UUID of the peer
very soon. When it gets that it compares the peer's UUIDs with its
own.

The problem is if our own current UUID changes between we sent it
and we got the peers current UUID. It can change because we are doing
a resync on an other connection.

When all that happens you see one end of a connection in
replication:WFBitMapS and the other end in eplication:Established