generalize write_on_node
Lots of tests use the dd command to write some data to a node, and they
all package it as their own "write_on_node" function. In order to reduce
code duplication and to make the tests more readable, introduce a
general "write" function on Volumes, which can be used to write an
arbitrary amount of random data to one or all volumes of a node.

Also introduce 2 helpers, "kib_to_blocks" and "mib_to_blocks", which can
be used to convert from more human-friendly KiB/MiB to the "number of
blocks" metric that dd expects.