new test: resync-reconciliation
This test configures a 3 node system, one primary. It writes a few blocks
on the primary, and tricks the secondaries to be a different positions
in the receive stream.
Then it wants to witness the reconciliation resync between the secondaries
after the primary disappeared. It assures resync direction, resync amount
and the data itself after the resync.