drbdtest.py: Replace the pyhton-level logfile catcher with one based on FDs
The problem with the previou approach was that the stderr of child
processes was not catched and put into the test.log.

I consider that as critical, since the test.log without stderr of
logscan is neary useless to understand the root cause why a test
fails.

As long we used only Jenkins the workaround was to look only at
the output catched by jenkins (it catches, both stdout and stderr).
Now since we started to use run.py I want that to be fixed where it
should be fixed.