testsuite: upgrade to Python 3
This is long overdue, but the test suite is now officially compatible
with Python 3.
The changes were mostly made using Python's "2to3" tool, which did most
syntactical changes automatically.
One manual change was that the custom "syslog" module is now called
"syslogd", because Python 3 already has a standard library module named
"syslog".

Note that this only applies to the "local" parts of the test suite, i.e.
everything that is run on the test coordinator. The parts on the test
nodes are still Python 2.7.