io-errors-diskless
This test simulates IO errors on diskfull nodes and verifies
that applications on diskless nodes are completely shielded
from those.

Up to DRBD-9.0.22 this test demonstrates a lost reference
on the DRBD module, and quite often a timer flooding the logs
with something like:
We did not send a P_BARRIER for 195358ms > ko-count (7) * timeout (60 * 0.1s); drbd kernel thread blocked?