new test: promote_while_connecting
We had a bug in the code where we took two locks not in the same
order in two threads. That happened when a new UUID was created
while the uuid_sem is hold during establishing a connection.

I think it is good to keep this test in the testsuite permanently,
it is just one of these many corner-cases that are otherwise not
regularily executed.