new test: resync-between-targets
This test creates a "due to diskstates" resync between two nodes that
have an outdates version of the data. After that both start a resync
with the owner of the new data, i.e. both become resync target to
that third node.
When the later resync (bringing the real data) finishes it sends
resync finished notifications to both sync sources. Up to recently
it was ignored on the node that became sync-source with an
inconsistent disk.

this bug was fixed in drbd-9.0.23