new test: resync-after-failover-to-dless
This tests prepares a cluster with two storage nodes and a diskless node.
One of the storage nodes is primary. Then it gets restated, and the
primary role is moved to the diskless.
Then the test makes sure the rebooting storage node connects to the other
storage node first and last to the diskless.

Up to drbd-9.0.23 it missed then to resync (again) from the other
storage node.