new test: resync-source-becomes-target2
This test is similar to the resync-source-becomes-target test. It
unveiled two bugs in DRBD (up to 9.0.24) which were not uncovered by
resync-source-becomes-target.
The main differences are that it resyncs a few hundred MBs and it sets
up DRBD's data structures so that select_best_resync_source() finds
the candidate before finding the other peer_device