virter: remove snapd from ubuntu images
It slows down startup by a few seconds (or significantly more seconds
on a slower system), and sometimes inexplicably hangs when shutting down,
so that virter runs into a timeout at the end of a provisioning step.

We don't need snapd in our test images, so get rid of it.