new test: resync-finish
Break the connection between two resynching nodes so that the final P_UUIDS110
packet does not get through. The resync-target sends that to signal the
sync-source node that it adopted the current-uuid of the sync-source.

This tests demonstrates that then that DRBD "forgets" to do the next resnyc
after such an event. Up to version (including) 9.0.24.