enable some stable tests
This enables some tests which are known to be very stable, so that the
total probability of passing a full test suite run is around 90%.

Some disabled tests are annotated to point out that there is something
wrong with them in particular.
"known broken" indicates that the test fails pretty much every time, and
is acknowledged to do so.
"known unstable" indicates a success rate of lower than 90%, which is
documented inline.