new test: resync-empty-after-primary-crash
This test creates a 3 node setup with quorum enables and simulates a
crash of the primary. Then a other node gets promoted, but not data
is written after promotion.
The resyncs after re-integration leave one node with an Outdated disk
frequently. 10 retries seem to be enough to trigger this bug.