write-during-resync: decrease delay
After the delay has been introduced, the write command sometimes took
too long. It typically took 20s on our test infrastructure, and
occasionally hit the timeout of 30s.

Reduce the delay to 50ms. This should still be comfortably enough to
ensure that the packets arrive at the peers in the correct order.