wait for initial state from events2
Otherwise early state changes may be missed. For instance, the test
'ref-count' does not wait for any events from the 'other' node until
resource.down() at the end. If the events2 stream is slow to start on
the 'other' node, the resource will already be down before the events
start. As a result, the test fails with:
Pattern 'destroy resource' does not match in lbtest-vm-55.test (line 1)