verify: fix integer division for python3
This fixes a spurious failure that occurred when 2 of the randomly
chosen differing blocks happened to be adjacent:

  File \"tests/verify\", line 118, in GetDifferenceFromDmesg"
    drbdtest.log(\"on node %s found change at sector %d (byte %d, 0x%x)\" % (node, sector, byte, byte))"
TypeError: %x format: an integer is required, not float"

The issue was that the operator used to adjust 'sector' was '/', causing
'sector' to become a float and thus 'byte' too.

Based on the number of blocks chosen and the range they are chosen from,
this situation would have occurred approximately 1 in every 300
attempts.