twopc_abort_connect: fix failures due to unexpected two-phase commits
Since DRBD commit
6caee0f43ba2 drbd: Fix two-primaries connect behaviour
this test sometimes failed (approximately 20% failures in my test
environment).

A failed connection now results in an aborted two-phase commit. Under
certain timing conditions, this caused node_b to have to retry the
two-phase commit when connecting to node_c. This caused the test to
fail with:
Exception: No two-phase commits expected at this moment

This is reasonable behavior from DRBD, very different from the two-phase
commit retry loop which occurred with drbd-9.0.24. Hence we adjust the
test to allow it.

It is sufficient just to wait for the connection with an appropriate
timeout, rather than checking the state of the two-phase commits. I have
verified that the test still fails with drbd-9.0.24.