diskless: fix order of writing files
This test was failing with more than 3 nodes. The issue was that we were
creating divergent data. We wrote newer data on the later nodes in the
file writing part. But then, in the re-integration part, while the later
nodes were disconnected, we wrote on the earlier nodes.

The test is still unstable with 5 nodes, but at least it passes
sometimes.