new test: resync-target-promote
This test simultes the initial full sync from a node that is forced to
consider itself as up-to-date.
While the resync runs one of the sync-targets promotes and demotes.

Up to (including) drbd-9.0.25 this led to an clear-of-bitmap while the
initial resync runs. It tiggers at least a log message; maybe other not
yet discovered negative effects.