resync-target-promote: replace log message check with down and up
The unexpected log messages still sometimes occur with drbd-9.0.26-0rc2.
However, the rest of the test is valid and important, so remove the log
checks for now.

This still fails fairly often on drbd-9.0.25 because the scenario
resulted in genuine data corruption with that version.

This approach has the advantage that we are not relying on logs which
are not really a stable API.