Revert "Improve the connect test"
This reverts commit 257f54adc36961f30159bdb3ed3efdb8fd53be74.

Although it is an improvement for both nodes to enter StandAlone state
when they fail to connect due to multiple primaries, the related changes
broke the "rr-conflict retry-connect" behavior. It is difficult to
support both, so we return to the old behavior when two primaries
connect. Remove the check for the new behavior.