switch-primaries: fix random failures
When "disconnect" is used, the test framework removes the connection
from its internal structures. Hence the "connect" afterwards did
nothing. The test only normally succeeded because there was a
'connection .* role:Primary' event in the events from the final write
loop or the iteration before. It failed with 2 nodes whenever all
iterations of the final loop, and the last iteration of the loop before,
chose the second node.

Basically we had a test which failed 1 in every 64 attempts.

This would have practically never occurred until
36110a6 reduce iterations of various slow tests