virter: get libvirt pool from environment
When using a libvirt pool other than "default" for virter, this image
pulling loop fails because it has "default" hardcoded.

Try a little bit harder to get the correct value by taking it from the
LIBVIRT_POOL variable if it is set.

This is still by no means perfect, since it does not consider the value
in ~/.config/virter/virter.toml, but it's a start.