two-primaries-reject: do not write between promoting and disconnecting
There is a genuine DRBD bug hidden here. But it is an obscure corner
case which will very rarely occur in any real situation. In the
meantime, until that bug is fixed, we still want the rest of the test to
work reliably. So we work around the issue. The extra write was never
really necessary to produce the scenarios that this test is intended to
verify.