limit the amount of memory which fio tries to allocate
Regardless of the number of jobs actually used, fio allocates memory for
a large number by default. Reduce this number so that fio does not try
to allocate so much memory.

This fixes an issue where fio was unable to start on CentOS 8 with
fio-3.19 in a VM with 1GiB memory. The error was of the form:
lbtest-vm-56.test: fio --section write -
[11s pause]
lbtest-vm-56.test: fio: job startup hung? exiting.