virter: allow separate drbd repository url
For the RHEL kernel puller, we want to test the latest released drbd
version. This means we need to get the drbd packages from our public
facing (customer) repositories, not the internal repositories used
for drbd testing.

However, we still want to fetch the extra packages (exxe, drbd-utils)
from the testing repositories.

Make it possible to specify a "DrbdRepositoryUrl" which will *only* be
used to fetch kmod-drbd. In order to be compatible with existing users,
just set DrbdRepositoryUrl = RepositoryUrl when it is unset, which is
essentially equivalent to the current behavior.