ci: simplify handling of extra arguments to run-test.sh
Previously, run-test.sh tried to pick out the --tests argument to avoid
passing it twice to vmshed. However, this logic was never used because
the tests file is set using --tests=... rather than as separate
arguments.  It works anyway because vmshed just takes the value from the
last occurrence of the argument.

Remove the unnecessary complexity.