get rid of drbdadm target helper
In an effort to eliminate the "global state" that is set via environment
variables, we want to pass the necessary environment variables only to
the commands that need them. The drbdadm helper is the biggest problem
here, because it requires PATH to be overriden.

Since this is an obfuscating piece of magic anyways, just get rid of it.
It comes at the expense of defining a drbdadm() function in the python
library, but I think it is much cleaner in the end.