new test: quorum-suspend-timeout
This tests configures network timeout (2 seconds) with a ko-count (2)
and on-no-quorum set to suspend-io.

Then it manouvers a primary into having open write-request with a
secondary which's disk is frozen.

First, the timeout has to trigger. After reconnect the secondarie's
disk is fine again. The test checks if the reconnect works, and
ensures last that the data on the backing devices is equal.

DRBD fails multiple aspects of this test as of 9.0.28