resync-reconciliation: allow primary_n to sync from either peer
After unblocking, primary_n may connect to the peers in either order and
hence sync from either. Rather than forcing it to connect to them in a
particular order, which would slow the test down even more, just allow
it to connect to either and be a bit more flexible in the events
checking.

We expect a sync whenever secondary_slow is promoted and not just in the
EARLY case.