virter: update virter and vmshed versions
This speeds up the pipeline because DHCP hosts are no longer added to
running networks.

It also opens up the possibility of using an extra network for the
"multi-path" test. Care will be required to ensure that the extra
interface gets an IP address on all test distributions.

We have to change the way that we add the additional labeled address
because the VMs are now assigned addresses in ranges like 10.224.1.0/24.