invalidate: correctly clear the DRBD device
fio rounds the size down to a multiple of the block size it is using.
Our DRBD device is a little smaller than 20MiB, so we can't use bs=1M to
write the whole device. Use bs=4K instead.

This never caused any actual problems because the entire device which is
used as our volume group is zeroed out in our test infrastructure.

When writing to the backing disk, as other tests do, it is OK to use
bs=1M because the backing disk size is a multiple of 1MiB.