virter: include the job ID in the elasticsearch index name
Since
4f8c095 virter: update vmshed version
the timestamp from when vmshed started instead of when it finished is
used. That makes it much more likely that our two scheduled stability
test runs have the same timestamp. Then the results from the second run
overwrite those from the first because the IDs are the same. This
occurred last night.

Avoid this problem by adding the job ID to the index name. Keep the date
in the index name because it might be useful for clearing up old
indices.