resync-source-becomes-target: do not wait for sync progress
The commit
81af2e4 resync-source-becomes-target: fix spurious timeouts
added waiting for sync progress events because the sync was sometimes
too slow. Unfortunately, this introduced a race condition because the
"after-resync-target" event and the "disk:UpToDate" event can be sent in
either order.

Since then some slow sync issues in DRBD have been fixed. Also, our
automated test infrastructure is now more performant and reliable. So it
is sufficient to just wait for an UpToDate disk. Remove the unnecessary
extra sync progress wait.