resync-source-becomes-target: wait for downstream sync in both variants
Even in variant 2 there is another downstream sync after the upstream
sync to get the latest data to "other_nodes[0]". Wait for this to
complete before continuing with the test. This prevents failures due to
two-phase commit timeouts when we take the resource down while this sync
is being negotiated.