resync-source-becomes-target: wait for peers to see UpToDate
To prevent race conditions when continuing the test after a sync is
supposed to be finished, wait for the sync source to see
"peer-disk:UpToDate" for the sync target.

Also remove the randomization of node order. This causes more trouble
than it is worth by making it difficult to debug the test.

Also rename the node and connection variables. Using explicit names
makes the test easier to understand than using list indexes -1, 0, 1.
The terms "forward" and "backward" are also not that clear to me, so
replace them with more explicit names.