resync-source-becomes-target: reduce sync rate for interrupted sync
This helps to ensure that variant 1 really tests what is intended.

Only 40M is synced from node_s0 to node_s1. Sync progress is reported
every 3s. So with "--c-max-rate=10M", we wait until 30M of the 40M sync
is complete before starting the upstream sync. It can easily happen that
the downstream sync is complete before the upstream sync starts. To
prevent this, simply reduce the max sync rate further.