replace exxe with ssh ControlMaster
exxe is a tool that enables efficient execution of commands on remote
machines, by "multiplexing" stdin, stdout, and stderr using only a
single SSH connection. It is quite good at that, but the added
dependency means more complexity and makes it more difficult to set up
test machines.

The problem with using plain SSH is that opening a brand new connection
for every command is just way too slow.

Fortunately, SSH has a built-in feature for connection multiplexing
called ControlMaster. When using this, a single connection is
established (and "remembered" through a local file) and future command
executions reuse that connection.

So, create a relatively small and simple Python wrapper for creating SSH
ControlMaster connections and use that instead of the external exxe
tool.