busy-resync: simplify test and make it fail with various fio versions
fio needs the "--max-jobs" parameter to work in our test environment.
Without this parameter it fails and does not perform any IO. This causes
the test to pass incorrectly. See:
accd449 limit the amount of memory which fio tries to allocate

In addition, fio-3.16 appears to have a bug affecting the combination of
the "size" and "time_based" options. See the comment in the code.

Also remove the unnecessary thread. We start fio in the background
anyway.