fix traffic control and make it more flexible
There was an issue with the way 'tc' was used which caused some packets
to be delayed which should not have been. Specifically, the netem qdisc
was attached to class 1 of the prio qdisc. However, the prio qdisc was
used with the default priomap, which sends packets with priority
'Interactive' to class 1. So these packets were sent to netem regardless
of whether they were matched by the filter or not. That is, they were
delayed regardless of the destination IP address.

Fix this by specifying priomap explicitly.

Also extract the 'tc' configuration into a re-usable component and
rework it so that multiple connections from be independently controlled.