virter: update virter and vmshed versions
Sometimes "virter vm rm" can take a long time and be canceled. If it is
canceled at the wrong point, the previous virter version could leave
orphaned volumes. These caused problems for later tests. The new virter
version fixes this so that volumes are never orphaned.

The new vmshed version changes the treatment of tests where the VMs fail
to start. These are now skipped instead of being counted as failed
tests.