drbdtest: remove unexpected c-min-rate configuration
This line has been present since the introduction of the python test
framework, but there is no clear reason for it. If a test is doing
application IO concurrently with resync, it should set or disable
c-min-rate explicitly. If this line is required otherwise, then there is
a DRBD bug.