convert options fields into properties
This way we can call touch_config() whenever the options are changed.
This removes a source of confusion when writing tests. It also means
that there is no difference between setting options before and after the
first write of the config file. So add_disk() is now free to write the
config file.

Also remove some unnecessary update_config() calls. This is done by
default whenever any command is run.