virter: separate the kernel that we build for from the one we test on
Sometimes the earlier kernels in a RHEL minor release series have bugs
that effect our tests. Also, we want to test on a system that is similar
to the ones used by real users.

The trigger for this change was an issue with 3.10.0-229.1.2.el7 which
caused it to sometimes fail to load drbd_transport_tcp with the error:
drbd_transport_tcp: Unknown symbol  (err 0)