resync_multiple-interrupt: resync a few blocks more
It looks like this test was handtuned to just work with the resync
as it was. I am working on the way the resync works. I guess I made
it a bit quicker. Now the tests needs to resync a bit more to
work with the quicker resync of DRBD.