resync-after-failover-to-dless: prevent concurrent demote and connect
Recent DRBD versions perform a two-phase commit for demotion. When we
try to do this while connecting, things get messy. In particular, the
demotion can fail with:
State change failed: (-19) Concurrent state changes detected and aborted

This is not an important aspect of the test, so avoid the issue by
waiting for the connection to be established before demoting.