peer-acks: keep DRBD open throughout test
DRBD will be changed to fsync and flush requests from upper layers when
the last opener is removed, instead of when demoted. This causes this
test as it previously was to fail because a peer-ack is sent when the
writing process closes the DRBD device file. Relax the test to allow
the new behavior.

In order to keep DRBD open, we use a filesystem. This causes additional
writes, so we have to adjust the checks for the number of out-of-sync
bits.