remove random choices from tests
The disadvantage of randomization is that it makes the tests more
complex. In particular, it is more complicated to investigate test
failures.

The advantage of randomization is that more cases are covered. This is a
basic form of fuzzing.

Experience shows that the cost of randomization outweighs the benefit.
Remove it.