drbdtest: make it possible to include generated resource configs
When developing tests, it is sometimes convenient to be able to
manipulate the test resources with drbdadm. Before this change, this was
a bit awkward. The way to use the resource config files was with:
$ drbdadm -c /var/lib/drbd-test/*/drbd.conf ...

Separate the "global" section from the rest of the config file to make
it possible to include the resource configs without having conflicting
"global" sections. With this include drbdadm can be used without "-c".

Also add such an include to the test images.