new-test: connect-abort
This test tries to establish the connection from A to B. C is already
connected to both.

   A   B
    \ /
     C

After the TCP connections are established, the network breaks while exchanging
the P_SYNC_PARAM89 packet.

In case the connect transaction lingers on node B, it will break the
promote on A.

This test fails up to drbd-9.1.7