new-test: test net namespaces
This test verifies that DRBD supports network namespaces. It only verifies
that the resource can be configured and that a connection can be established.

In order to test net namespaces, the test setup creates a new network namespace
and moves the secondary network interface (which must exist and have an ip
address assigned) to the new namespace. All later commands (drbdadm, etc..)
are executed inside that namespace.