new-test: diskless-connect-older
Consider a cluster with a primary diskless node, that has connections
with two diskfull nodes. One of the diskful nodes leaves the cluster
for some time. When it comes back it connects to the diskless node first.
This connection should be accepted by the primary diskless, and the
joining node's disk needs to be downgraded below UpToDate at this
moment.
As the connect logic depends on node ids, this test-program does the
test twice (with both variations of node-ids)

Up to drbd-9.1.7 the primary diskless simply rejected the connect attempt.