ci: add ability to rebuild drbd
When we rebuild the test images to update kernels, we cannot run the
test suite right away because the drbd modules for the new kernels have
not yet been built.

Currently, we manually trigger the drbd pipeline so that we can verify
that the new images work. This is tedious, so automate it in a pipeiline
step.