ci: use correct hostname based on DRBD version
When testing with DRBD 8.4, drbdadm takes the nodes "short hostname" as
its hostname, which differs from 9.0 where the fully qualified domain
name is used.

This creates a problem where drbdadm does not find any resources because
the hostname in the DRBD config is fully qualified while the nodes
hostname is not.

So, when running DRBD < 9.0.0, use the "short hostname" as our hostname
in the test framework instead.