always include all connections in Node.connections
Prior to this change, Node.connections only included those connections
which had not been disconnected. This applied by extension also to
Nodes.connections and Resource.connections. This behavior is
counter-intuitive. Simplify the code so that all connections are
included regardless of whether we have disconnected them.

The last commit adjusted those few tests that did rely on the old
behavior. They no longer rely on it, so this change should not break any
tests.

This change is nevertheless a functional change. Some tests now wait for
events that were intended, but were ignored due to the counter-intuitive
behavior. The most widespread impact is on the new_peer and new_path
methods. These now wait for the respective create events even when the
peer has been disconnected.