raw storage: reserve space for partitions to grow
some tests with multiple volumes still want to resize the backing disks.
In order to do that with partitions we need to reserve enough space for the
partition, so that we can expand from the existing start sector.

To do that, we introduce a new parameter that specifies the maximum size
we expect the disk to grow, and use that when initially creating the partition
to ensure enough space is available after the initial end to extend it later.