new-test: test quorum when adding nodes to cluster
Add tests that ensure quorum rules stay consistent when adding nodes
to an existing cluster, specifically a degraded cluster.

First test checks that adding a new diskfull node to a cluster causes
a brief loss of quorum until the new disk is connected to the remaining
quorate partition.

Second test checks that adding a new diskfull node to a cluster causes
a switch in quorate partitions when the new disk is connected only to
the previously isolated partition.

Third and fourth test ensures adding a diskless does not influence
current quorum.