ci: disable CI tests which are unstable with multiple CPUs
It is not yet clear whether the tests are unstable due to issues with
the tests themselves, issues in DRBD, or the increased competition for
host CPU resources.

Disable them for now to keep our CI tests sufficiently stable. They can
be investigated and re-enabled once the instability has been fixed.