drbdtest: make fencing mode per-node
This allows us to put the "fencing" parameter either in the "disk" or
"net" sections, depending on the drbd version being used.

Use the new property to set the fencing mode in the `fencing` and
`suspend-io` tests. This should make them compatible with drbd 8.4.

Also disable some test cases in `fencing` that do not work with 8.4.