ci: adjust CI tests
Unfortunately the success rates of some of the tests have fallen over
time to the point where the CI runs are no longer reliable enough.
Adjust the set of CI tests to restore trust in the CI runs.

Note that fencing-5 has been kept even though it is slightly below the
threshold, in order to ensure that we keep some non-trivial 5 node
testing.