new test: quorum-check-quorumless
In a three node cluster with C diskless:

     C
    / \
   A---B

When B is primary and gets isolated from the other two nodes. Then the
remaining partition of A & C know that B can not have quorum. When a
bit later C disconnects from A, A can still be sure that B (and C) can
not re-gain quorum. Therefore A still has quorum.

The implementation relies on remembering which nodes are "quorumless".

DRBD-9.1.13 and 9.2.2 are the first releases that do that.