virter: uninstall bcache-tools if it is installed
When testing drbd 8.4 with Ubuntu, udev sometimes probes the "device
under test", apparently to check for a bcache on the disk. While drbd 9
has a workaround for this built in, drbd 8 will fail to demote the
device in this case.

So, to hopefully have nothing access the device while the test is
running, uninstall bcache-tools if it exists.

This does not seem to exist on AlmaLinux 8, so we should be fine for
RHEL-based distros:

$ yum whatprovides /lib/udev/probe-bcache
Error: No Matches found