add disk volume classes
This is useful for storing some additional data for each disk volume.
For instance we can store the raw volume partition number to avoid the
need to collect it whenever we need it.

This will also help with adding an in-memory pool and a disk delay
layer. These are useful for endurance tests.