ci: add job to build DRBD for a given ref before running tests
This is useful for running custom tests, such as those from a branch or
with special parameters, against DRBD from a branch. Without this job,
the developer must build DRBD, wait for the build to complete, construct
the appropriate package version and then start the tests pipeline. Now
this can all be done with one command.