quorum-check-quorumless: Adapt test for DRBD improvements
When DRBD learns to propagate the reachable_nodes mask with two-phase
commits the quorum-check-quorumless needs to be addapted to that.

Now it verifies that:
Nodes A, B, and C with ascending node IDs.

        C       | C is diskless
       / \      | A is Secondary and has a backing disk
      A---B     | B is Secondary, having a backing disk

Step 1: A does a graceful disconnect from B

        C       | C is diskless and has quorum
       / \      | A is Secondary with disk and has quorum
      A   B     | B is Secondary with disk and has quorum

Step 2: A does a graceful disconnect from C

        C       | C is with quorum since it is part of 2 out of 3
         \      | A is **without** quorum
      A   B     | B is with quorum since it is part of 2 out of 3

Step 3: C does a graceful disconnect from B

        C       | C is without quorum
                | A is without quorum
      A   B     | B is ...

The old way of verifying that a connect propagates is no longer
relevant.

Also the may-promote test needs a minor change.